UBUNTU: [Config] re-enable OSS support and blacklist
We are seeing requests for the OSS support to be reenabled, it seems that
it is needed for old games.  We do not want it loading by default so use
the new blacklist support to suppress autoloading.  This should bring us
into line with other distros where manual loading is required.

BugLink: http://bugs.launchpad.net/bugs/1434842
Signed-off-by: Andy Whitcroft <apw@canonical.com>