UBUNTU: [Config] switch squashfs to single threaded decode
There is some issue with squashfs decoding when done in a multi-threaded
manner which leads to large memory consumption.  Either we have a leak
or more probabally we have pathalogical case leading to horrible internal
fragmentation.  For the moment turn it off while it can be investigated.

BugLink: http://bugs.launchpad.net/bugs/1636847
Signed-off-by: Andy Whitcroft <apw@canonical.com>
Acked-by: Brad Figg <brad.figg@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>