UBUNTU: [Config] Fix up CONFIG_I2C_SLAVE values
The annotations file says this should be =n on all architectures
except for armhf, but it is currently =y on several others. This
is because it's being selected by CONFIG_I2C_EMEV2, which is only
really relevant for armhf, and CONFIG_I2C_RCAR, which is only
relevant on armhf and arm64. Disable CONFIG_I2C_EMEV2 and
CONFIG_I2C_SLAVE for all other architectures, and update the
annotations file entry for CONFIG_I2C_EMEV2 accordingly.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>