UBUNTU: SAUCE: aufs -- Add flags argument to aufs_rename()
Starting with Linux 4.9-rc1 the rename2 inode operation has
replaced the rename op, so filesystem rename implementations
require an extra flags argument. Add the argument to fix the
FTBFS with Linux 4.9, but since aufs doesn't support any of the
flags return an error if it is non-zero.

Fixes FTBFS with 4.9 kernels.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>