UBUNTU: SAUCE: selftests/powerpc: Use snprintf to construct DSCR sysfs interface paths
BugLink: http://bugs.launchpad.net/bugs/1710904

Currently sprintf is used, and while paths should never exceed
the size of the buffer it is theoretically possible since
dirent.d_name is 256 bytes. As a result this trips
-Wformat-overflow, and since the test is built with -Wall -Werror
the causes the build to fail. Switch to using snprintf and skip
any paths which are too long for the filename buffer.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>